Updated hit approach
Using beziere curve with tracked argets. The arc of the curve need to be hit though which is not happening at the moment (could emphasise this is the beziere somehow). This all needs to be controlled so the club does not then hit into the player and acute angles. but good progress on making it look like a swing